requirements.txt
typer>=0.3.0
mediapipe>=0.8.3.0
pillow>=8.1.0
opencv-python>=4.5.1.0
aiortc>=1.2.0
